GoogleScholarAgent/google_scholar_02/requirements.txt
google-adk
google-cloud-aiplatform
absl-py
python-dotenv
requests

google-cloud-aiplatform[adk,agent_engines]
beautifulsoup4
lxml
google-search-results
pandas